requirements.txt
docker
coloredlogs
PyYAML
